refactor(commands): mod active_environments
Split everything related to active environments out to a separate module
because `commands` is getting very big and hard to navigate.
